Document OpenAPT solution in integrations
OpenAPT is a tool developed by [ALLOcloud](https://allocloud.com) team.
It's a file format which allows user to describe APT repositories
specifications with a JSON/YAML file. It's also an implementation in
Python which automate aptly based on these specifications.

The goal is to document the layout of APT repositories and automate their
build with aptly commands. Flow and keywords are inspired by core entities
from aptly.